// src/app/interfaces/update-patient.ts
export interface UpdatePatient {
  idPaciente: number;
  codigoUbigeo: string;
  apellidoMaterno: string;
  idSexo: number;
  direccion: string;
  lugarNacimiento: string;
  codigoAsegurado: string;
  numeroDocumento: string;
  idTipoDocumento: number;
  fechaNacimiento: Date;
  nombres: string;
  apellidoPaterno: string;
}